+++ fontes/primitivas/dialetos/visualg.ts
@@ -1,3 +1,12 @@
+export const estruturasDados: {nome: string, descricao?: string, documentacao: string, exemploCodigo?: string}[] = [
+    {
+        nome: 'vetor',
+        descricao: 'Variável que pode ter uma ou duas dimensões. Um vetor de duas dimensões também é chamado de matriz.',
+        documentacao: '# Declaração `vetor`\nVariável que pode ter uma ou duas dimensões. Um vetor de duas dimensões também é chamado de matriz.',
+        exemploCodigo: 'var meuVetor: vetor[1..5] de inteiro'
+    }
+];
+
 export const primitivasEntradaSaidaVisuAlg: {nome: string, descricao?: string, documentacao: string, exemploCodigo?: string}[] = [
     {
         nome: 'escreva',